wip: initial work for reorganizing modules
