+++ cool_seq_tool/__init__.py
@@ -1,5 +1,4 @@
 """The cool_seq_tool package"""
-from os import environ
 from pathlib import Path
 import logging
 
@@ -13,5 +12,3 @@ logger = logging.getLogger("cool_seq_tool")
 logger.setLevel(logging.DEBUG)
 
 LOG_FN = "cool_seq_tool.log"
-
-from .app import CoolSeqTool  # noqa: E402, F401, I202
