--- cool_seq_tool/data_sources/__init__.py
@@ -1,7 +1,0 @@
-"""Module for data sources"""
-from .seqrepo_access import SeqRepoAccess
-from .mane_transcript_mappings import MANETranscriptMappings
-from .transcript_mappings import TranscriptMappings
-from .uta_database import UTADatabase
-from .mane_transcript import MANETranscript
-from .alignment_mapper import AlignmentMapper
