+++ cool_seq_tool/handlers/__init__.py
@@ -1,0 +1,2 @@
+"""Module for extending clients"""
+from .seqrepo_access import SeqRepoAccess
