+++ cool_seq_tool/mappers/__init__.py
@@ -1,0 +1,4 @@
+"""Module for mapping data"""
+from .alignment import AlignmentMapper
+from .mane_transcript import MANETranscript
+from .exon_genomic_coords import ExonGenomicCoordsMapper
