+++ cool_seq_tool/sources/__init__.py
@@ -1,0 +1,4 @@
+"""Module for providing basic acquisition/setup for the various resources"""
+from .mane_transcript_mappings import MANETranscriptMappings
+from .transcript_mappings import TranscriptMappings
+from .uta_database import UTADatabase
